+++ src/java/main/serviceLayer/Service.java
@@ -96,7 +96,7 @@ public class Service implements IService {
     }
 
     @Override
-    public Response addItemToShoppingCart(ItemFacade itemToInsert, int amount, String shopName, String visitorName) {
+    public Response addItemToShoppingCart(ItemFacade itemToInsert, double amount, String shopName, String visitorName) {
         return null;
     }
